Extract convergence checks function from `SamplerReport`
